+++ src/xcode/ENA/ENA/Source/Scenes/AppInformation/AppInformationImprintTest.swift
@@ -1,10 +1,20 @@
 //
-//  AppInformationImprintTest.swift
-//  ENA
+// Corona-Warn-App
 //
-//  Created by Vogel, Andreas on 01.09.20.
-//  Copyright © 2020 SAP SE. All rights reserved.
+// SAP SE and all other contributors
+// copyright owners license this file to you under the Apache
+// License, Version 2.0 (the "License"); you may not use this
+// file except in compliance with the License.
+// You may obtain a copy of the License at
 //
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 
 import XCTest
 @testable import ENA
@@ -16,7 +26,7 @@ class AppInformationImprintTest: XCTestCase {
 			.imprint: (
 				text: AppStrings.AppInformation.imprintNavigation,
 				accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.imprintNavigation,
-				action: .push(model: appInformationImprintModel.dynamicTable, withTitle:  AppStrings.AppInformation.imprintNavigation)
+				action: .push(model: AppInformationImprintModel.dynamicTable, withTitle:  AppStrings.AppInformation.imprintNavigation)
 			)
 		]
 		
@@ -25,10 +35,10 @@ class AppInformationImprintTest: XCTestCase {
 		let key = model.first?.key
 		XCTAssert(key == .imprint)
 		
-		let dynamicTable = appInformationImprintModel.dynamicTable
+		let dynamicTable = AppInformationImprintModel.dynamicTable
 		XCTAssertTrue(dynamicTable.numberOfSection == 1)
 		
-		let section = appInformationImprintModel.dynamicTable.section(0)
+		let section = AppInformationImprintModel.dynamicTable.section(0)
 		XCTAssertNotNil(section)
 		let numberOfCells = section.cells.count
 		
@@ -42,7 +52,7 @@ class AppInformationImprintTest: XCTestCase {
 	}
 	
 	func testContactForm() {
-		let cellCollection = appInformationImprintModel.contactForms()
+		let cellCollection = AppInformationImprintModel.contactForms()
 		let numberOfCells = cellCollection.count
 		
 		let localization = Bundle.main.preferredLocalizations.first
@@ -52,4 +62,5 @@ class AppInformationImprintTest: XCTestCase {
 			XCTAssertTrue(numberOfCells == 2) //else
 		}
 	}
+	
 }
