+++ src/xcode/ENA/ENA/Source/Scenes/AppInformation/AppInformationViewController+DynamicTableViewModel.swift
@@ -1,3 +1,4 @@
+//
 // Corona-Warn-App
 //
 // SAP SE and all other contributors
@@ -23,7 +24,7 @@ extension AppInformationViewController {
 		.about: (
 			text: AppStrings.AppInformation.aboutNavigation,
 			accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.aboutNavigation,
-			action: .push(model: aboutModel, withTitle:  AppStrings.AppInformation.aboutNavigation)
+			action: .push(model: AppInformationModel.aboutModel, withTitle:  AppStrings.AppInformation.aboutNavigation)
 		),
 		.faq: (
 			text: AppStrings.AppInformation.faqNavigation,
@@ -33,12 +34,12 @@ extension AppInformationViewController {
 		.terms: (
 			text: AppStrings.AppInformation.termsTitle,
 			accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.termsNavigation,
-			action: .push(model: termsModel, withTitle:  AppStrings.AppInformation.termsNavigation)
+			action: .push(model: AppInformationModel.termsModel, withTitle:  AppStrings.AppInformation.termsNavigation)
 		),
 		.privacy: (
 			text: AppStrings.AppInformation.privacyNavigation,
 			accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.privacyNavigation,
-			action: .push(model: privacyModel, withTitle:  AppStrings.AppInformation.privacyNavigation)
+			action: .push(model: AppInformationModel.privacyModel, withTitle:  AppStrings.AppInformation.privacyNavigation)
 		),
 		.legal: (
 			text: AppStrings.AppInformation.legalNavigation,
@@ -48,97 +49,12 @@ extension AppInformationViewController {
 		.contact: (
 			text: AppStrings.AppInformation.contactNavigation,
 			accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.contactNavigation,
-			action: .push(model: contactModel, withTitle:  AppStrings.AppInformation.contactNavigation)
+			action: .push(model: AppInformationModel.contactModel, withTitle:  AppStrings.AppInformation.contactNavigation)
 		),
 		.imprint: (
 			text: AppStrings.AppInformation.imprintNavigation,
 			accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.imprintNavigation,
-			action: .push(model: appInformationImprintModel.dynamicTable, withTitle:  AppStrings.AppInformation.imprintNavigation)
+			action: .push(model: AppInformationImprintModel.dynamicTable, withTitle:  AppStrings.AppInformation.imprintNavigation)
 		)
 	]
-}
-
-extension AppInformationViewController {
-	private static let aboutModel = DynamicTableViewModel([
-		.section(
-			header: .image(UIImage(named: "Illu_AppInfo_UeberApp"),
-						   accessibilityLabel: AppStrings.AppInformation.aboutImageDescription,
-						   accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.aboutImageDescription,
-						   height: 230),
-			cells: [
-				.title2(text: AppStrings.AppInformation.aboutTitle,
-						accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.aboutTitle),
-				.headline(text: AppStrings.AppInformation.aboutDescription,
-						  accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.aboutDescription),
-				.subheadline(text: AppStrings.AppInformation.aboutText,
-							 accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.aboutText),
-				.bodyWithoutTopInset(text: germanLinkText(), style: .linkTextView(AppStrings.AppInformation.aboutLinkText, .subheadline), accessibilityIdentifier: AppStrings.AppInformation.aboutLinkText)
-			]
-		)
-	])
-
-	private static let contactModel = DynamicTableViewModel([
-		.section(
-			header: .image(UIImage(named: "Illu_Appinfo_Kontakt"),
-						   accessibilityLabel: AppStrings.AppInformation.contactImageDescription,
-						   accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.contactImageDescription,
-						   height: 230),
-			cells: [
-				.title2(text: AppStrings.AppInformation.contactTitle,
-						accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.contactTitle),
-				.body(text: AppStrings.AppInformation.contactDescription,
-					  accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.contactDescription),
-				.headline(text: AppStrings.AppInformation.contactHotlineTitle,
-						  accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.contactHotlineTitle),
-				.phone(text: AppStrings.AppInformation.contactHotlineText, number: AppStrings.AppInformation.contactHotlineNumber,
-					   accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.contactHotlineText),
-				.footnote(text: AppStrings.AppInformation.contactHotlineDescription,
-						  accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.contactHotlineDescription),
-				.footnote(text: AppStrings.AppInformation.contactHotlineTerms,
-						  accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.contactHotlineTerms)
-			]
-		)
-	])
-
-
-	private static let privacyModel = DynamicTableViewModel([
-		.section(
-			header: .image(
-				UIImage(named: "Illu_Appinfo_Datenschutz"),
-				accessibilityLabel: AppStrings.AppInformation.privacyImageDescription,
-				accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.privacyImageDescription,
-				height: 230
-			),
-			cells: [
-				.title2(
-					text: AppStrings.AppInformation.privacyTitle,
-					accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.privacyTitle),
-				.html(url: Bundle.main.url(forResource: "privacy-policy", withExtension: "html"))
-			]
-		)
-	])
-
-	private static let termsModel = DynamicTableViewModel([
-		.section(
-			header: .image(UIImage(named: "Illu_Appinfo_Nutzungsbedingungen"),
-						   accessibilityLabel: AppStrings.AppInformation.termsImageDescription,
-						   accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.termsImageDescription,
-						   height: 230),
-			cells: [
-				.title2(
-					text: AppStrings.AppInformation.termsTitle,
-					accessibilityIdentifier: AccessibilityIdentifiers.AppInformation.termsTitle),
-				.html(url: Bundle.main.url(forResource: "usage", withExtension: "html"))
-			]
-		)
-	])
-}
-
-// return a link if language is German; else return an empty string
-private func germanLinkText() -> String {
-	if Bundle.main.preferredLocalizations.first == "de" {
-		return AppStrings.AppInformation.aboutLink
-	} else {
-		return ""
-	}
 }
